게시물 삭제(수정)

+++ src/main/java/com/karis/exam2/user/controller/UserController.java
@@ -48,6 +48,9 @@ public class UserController {
     @RequestMapping("doDelete")
     @ResponseBody
     public String showDelete(long id) {
+        if(articleRepository.existsById(id) == false){
+            return "%d번 게시물은 없습니다.".formatted(id);
+        }
         articleRepository.deleteById(id);
         return "%d번 게시물이 삭제되었습니다.".formatted(id);
     }
